refactor : #19
